refactored players controls
